msm: ipa: Fix the handling of default IPA header
Default IPA header is added or deleted from the driver
directly and not by user space application. This change
prevents adding/deleting it from user application which
may cause inconsistencies in the driver. Also the change
fixes the header reset function to skip on the correct
default header.

Bug: 72957269
CRs-fixed: 2151146
Signed-off-by: Ghanim Fodi <gfodi@codeaurora.org>

Conflicts:
	drivers/platform/msm/ipa/ipa_v2/ipa_hdr.c
	drivers/platform/msm/ipa/ipa_v3/ipa_hdr.c